Update prom check (#3699)
* Refactor prom check using mixin (port of https://github.com/DataDog/integrations-core/pull/1006)

* Add wildcard support
